add end2end tests
